pyproject.toml
[project]
name = "bme-scraper"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "load-dotenv>=0.1.0",
    "python-dotenv>=1.0.1",
    "selenium>=4.29.0",
    "webdriver-manager>=4.0.2",
]
